+++ src/main/java/ru/yandex/practicum/filmorate/storage/UserStorage.java
@@ -1,9 +1,7 @@
 package ru.yandex.practicum.filmorate.storage;
 
 import ru.yandex.practicum.filmorate.model.User;
-
 import java.util.List;
-import java.util.Map;
 
 public interface UserStorage {
 
@@ -11,7 +9,7 @@ public interface UserStorage {
 
     public void put(User user);
 
-    public User get(int id);
+    public User get(Long id);
 
     public User update(User user);
 }
